Separating schemas and routes for the players data API
