+++ nbaspa_app/io/players/__init__.py
@@ -1,0 +1,1 @@
+"""Import path."""
